requirements.txt
PyMuPDF==1.23.6
python-dateutil==2.8.2
Flask==3.0.0